+++ MediatorSetup/startup_without_elasticsearch.sh
@@ -3,14 +3,14 @@ set -e
 
 install_grafana() {
 	if [ -x "$(command -v yum)" ]; then
-		yum install -y https://dl.grafana.com/oss/release/grafana-11.0.0-1.x86_64.rpm
+		yum install -y https://dl.grafana.com/oss/release/grafana-11.2.0-1.x86_64.rpm
 	elif [ -x "$(command -v apt-get)" ]; then
-		apt-get install -y adduser libfontconfig1 musl
-		wget https://dl.grafana.com/oss/release/grafana_11.0.0_amd64.deb
-		dpkg -i grafana_11.0.0_amd64.deb
+    		apt-get install -y adduser libfontconfig1 musl
+	        wget https://dl.grafana.com/oss/release/grafana_11.2.0_amd64.deb
+        	dpkg -i grafana_11.2.0_amd64.deb
 	elif [ -x "$(command -v rpm)" ]; then
-		wget https://dl.grafana.com/oss/release/grafana-11.0.0-1.x86_64.rpm
-		rpm -Uvh grafana-11.0.0-1.x86_64.rpm
+		wget https://dl.grafana.com/oss/release/grafana-11.2.0-1.x86_64.rpm
+		rpm -Uvh grafana-11.2.0-1.x86_64.rpm
 	else
 		echo "Error can't install Grafana, please install it manually and re-run the script."
 		exit 1;